Add initial support for Redfish physical provider
This commit contains the bare minimum support for the API to be
used by the integration tests, which test the Redfish type of
the physical infrastructure providers [1].
The API implementation currently does not yet make any actual
Redfish client calls. Instead, it encodes the state as we expect
it from the mock-up server [2].
In this commit, we only validate the number of servers disovered
by the provider.

[1] https://github.com/ManageIQ/manageiq-providers-redfish
[2] https://github.com/xlab-si/redfish-recordings